requirements.txt
PyYAML>=6.0,<7.0
requests>=2.25.1,<3.0
